move awsclient into _types
