+++ pkg/types/certificaterequest_types.go
@@ -22,15 +22,24 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-)
+	"github.com/openshift/certman-operator/pkg/awsclient"
 
-// ReconcileCertificateRequest reconciles a CertificateRequest object
-type ReconcileCertificateRequest struct {
-	client           client.Client
-	scheme           *runtime.Scheme
-	recorder         record.EventRecorder
-	awsClientBuilder func(kubeClient client.Client, secretName, namespace, region string) (awsclient.Client, error)
-}
+	"context"
+	"fmt"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/types"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/openshift/certman-operator/pkg/types"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/route53"
+	"github.com/aws/aws-sdk-go/service/route53/route53iface"
+)
 
 // CertificateRequestSpec defines the desired state of CertificateRequest
 // +k8s:openapi-gen=true
@@ -175,4 +184,144 @@ const (
 func init() {
 	// Register adds its arguments (objects) to SchemeBuilder so they can be added to a Scheme.
 	SchemeBuilder.Register(&CertificateRequest{}, &CertificateRequestList{})
+}
+
+// ReconcileCertificateRequest reconciles a CertificateRequest object
+type ReconcileCertificateRequest struct {
+	client           client.Client
+	scheme           *runtime.Scheme
+	recorder         record.EventRecorder
+	awsClientBuilder func(kubeClient client.Client, secretName, namespace, region string) (awsclient.Client, error)
+}
+
+///////////////////////
+// AWS client types  //
+///////////////////////
+
+const (
+	awsCredsSecretIDKey     = "aws_access_key_id"
+	awsCredsSecretAccessKey = "aws_secret_access_key"
+)
+
+// Client is a wrapper object for actual AWS SDK clients to allow for easier testing.
+type Client interface {
+	// Route53 client
+	CreateHostedZone(input *route53.CreateHostedZoneInput) (*route53.CreateHostedZoneOutput, error)
+	DeleteHostedZone(input *route53.DeleteHostedZoneInput) (*route53.DeleteHostedZoneOutput, error)
+	ListHostedZones(input *route53.ListHostedZonesInput) (*route53.ListHostedZonesOutput, error)
+	GetHostedZone(*route53.GetHostedZoneInput) (*route53.GetHostedZoneOutput, error)
+	ChangeResourceRecordSets(*route53.ChangeResourceRecordSetsInput) (*route53.ChangeResourceRecordSetsOutput, error)
+	ListResourceRecordSets(*route53.ListResourceRecordSetsInput) (*route53.ListResourceRecordSetsOutput, error)
+}
+
+// awsClient implements the Client interface
+type awsClient struct {
+	route53Client route53iface.Route53API
+}
+
+// TestAuth ensures that AWS credentials are present,
+// and that they contain the permisionssion required
+// for running this operator.
+func (c *awsClient) TestAuth(cr *CertificateRequest, r *ReconcileCertificateRequest) error {
+	platformSecretName := cr.Spec.PlatformSecrets.AWS.Credentials.Name
+
+		awscreds := &corev1.Secret{}
+		err := cr.client.Get(context.TODO(), types.NamespacedName{Name: platformSecretName, Namespace: request.Namespace}, awscreds)
+	    if err != nil {
+		reqLogger.Info("platformSecrets were not found. Unable to search for certificates in cloud provider platform")
+			return reconcile.Result{}, credserr
+		}
+	// Ensure that platform Secret can authenticate to AWS.
+	   r53svc, err := r.getAwsClient(cr)
+		if err != nil {
+reqLogger.Error(err, err.Error())
+}
+	
+hostedZoneOutput, err := r53svc.ListHostedZones(&route53.ListHostedZonesInput{})
+if err != nil {
+reqLogger.Info("platformSecrets are either invalid, or don't have permission to list Route53 HostedZones")
+reqLogger.Error(err, err.Error())
+return reconcile.Result{}, err
+}
+	
+println("Successfully authenticated with cloudprovider. Hosted zones found:")
+println(hostedZoneOutput)
+	
+	return nil
+}
+
+func (c *awsClient) ListHostedZones(input *route53.ListHostedZonesInput) (*route53.ListHostedZonesOutput, error) {
+	return c.route53Client.ListHostedZones(input)
+}
+
+func (c *awsClient) CreateHostedZone(input *route53.CreateHostedZoneInput) (*route53.CreateHostedZoneOutput, error) {
+	return c.route53Client.CreateHostedZone(input)
+}
+
+func (c *awsClient) DeleteHostedZone(input *route53.DeleteHostedZoneInput) (*route53.DeleteHostedZoneOutput, error) {
+	return c.route53Client.DeleteHostedZone(input)
+}
+
+func (c *awsClient) GetHostedZone(input *route53.GetHostedZoneInput) (*route53.GetHostedZoneOutput, error) {
+	return c.route53Client.GetHostedZone(input)
+}
+
+func (c *awsClient) ChangeResourceRecordSets(input *route53.ChangeResourceRecordSetsInput) (*route53.ChangeResourceRecordSetsOutput, error) {
+	return c.route53Client.ChangeResourceRecordSets(input)
+}
+
+func (c *awsClient) ListResourceRecordSets(input *route53.ListResourceRecordSetsInput) (*route53.ListResourceRecordSetsOutput, error) {
+	return c.route53Client.ListResourceRecordSets(input)
+}
+
+// NewClient returns an awsclient.Client object to the caller. If NewClient is passed a non-null
+// secretName, an attempt to retrieve the secret from the namespace argument will be performed.
+// AWS credentials are returned as these secrets and a new session is initiated prior to returning
+// a route53Client. If secrets fail to return, the IAM role of the masters is used to create a
+// new session for the client.
+func NewClient(kubeClient client.Client, secretName, namespace, region string) (Client, error) {
+
+	awsConfig := &aws.Config{Region: aws.String(region)}
+
+	if secretName != "" {
+		secret := &corev1.Secret{}
+		err := kubeClient.Get(context.TODO(),
+			types.NamespacedName{
+				Name:      secretName,
+				Namespace: namespace,
+			},
+			secret)
+
+		if err != nil {
+			return nil, err
+		}
+
+		accessKeyID, ok := secret.Data[awsCredsSecretIDKey]
+		if !ok {
+			return nil, fmt.Errorf("AWS credentials secret %v did not contain key %v",
+				secretName, awsCredsSecretIDKey)
+		}
+
+		secretAccessKey, ok := secret.Data[awsCredsSecretAccessKey]
+		if !ok {
+			return nil, fmt.Errorf("AWS credentials secret %v did not contain key %v",
+				secretName, awsCredsSecretAccessKey)
+		}
+
+		awsConfig.Credentials = credentials.NewStaticCredentials(
+			strings.Trim(string(accessKeyID), "\n"),
+			strings.Trim(string(secretAccessKey), "\n"),
+			"",
+		)
+	}
+
+	//// Otherwise default to relying on the IAM role of the masters where the actuator is running:
+	s, err := session.NewSession(awsConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return &awsClient{
+		route53Client: route53.New(s),
+	}, nil
 }
